refactor(dust-hive): extract TUI script to dedicated module + allow initial service
- Move service-logs-tui.sh generation from open.ts to lib/scripts.ts
- Add optional [service] argument to start TUI on a specific service
- Validate service name in logsCommand when using -i with a service

The script is now organized in logical sections (header, utilities, UI,
tail management, service functions, main loop) making it more maintainable.
