+++ x/henry/dust-hive/src/commands/logs.ts
@@ -1,8 +1,8 @@
 import { withEnvironment } from "../lib/commands";
 import { getLogPath } from "../lib/paths";
 import { CommandError, Err, Ok } from "../lib/result";
+import { ensureServiceLogsTui } from "../lib/scripts";
 import { ALL_SERVICES, type ServiceName, isServiceName } from "../lib/services";
-import { ensureServiceLogsTui } from "./open";
 
 interface LogsOptions {
   follow?: boolean;
@@ -17,8 +17,15 @@ export const logsCommand = withEnvironment(
 
     // Interactive mode: launch the TUI
     if (interactive) {
+      // Validate service name if specified
+      if (serviceArg && !isServiceName(serviceArg)) {
+        console.log(`\nServices: ${ALL_SERVICES.join(", ")}`);
+        return Err(new CommandError(`Unknown service '${serviceArg}'`));
+      }
       const scriptPath = await ensureServiceLogsTui();
-      const proc = Bun.spawn([scriptPath, env.name], {
+      // If a service is specified, start the TUI on that service
+      const args = serviceArg ? [scriptPath, env.name, serviceArg] : [scriptPath, env.name];
+      const proc = Bun.spawn(args, {
         stdin: "inherit",
         stdout: "inherit",
         stderr: "inherit",
